refactor: extract EncodingDiagram and add an optional bit diff mask
